requirements.txt
GitPython>=0.1.7
argparse>=1.2.1
